rename instances to derived_optimizers
